+++ terrawrap/utils/config.py
@@ -9,7 +9,7 @@ import jsons
 import yaml
 from ssm_cache import SSMParameterGroup
 
-from terrawrap.exceptions import NotTerraformConfigDirectory
+from terrawrap.exceptions import NotTerraformConfigDirectory, NoDependency
 from terrawrap.models.wrapper_config import (
     WrapperConfig,
     AbstractEnvVarConfig,
@@ -22,6 +22,7 @@ from terrawrap.utils.path import get_absolute_path, calc_repo_path
 
 DEFAULT_REGION = 'us-west-2'
 SSM_ENVVAR_CACHE = SSMParameterGroup(max_age=600)
+TF_WRAP_FILE = ".tf_wrapper"
 
 
 def find_variable_files(path: str) -> List[str]:
@@ -62,7 +63,7 @@ def find_wrapper_config_files(path: str) -> List[str]:
     for element in elements:
         cur_path = os.path.join(cur_path, element)
         for file in os.listdir(cur_path):
-            if file.endswith(".tf_wrapper") or file.endswith(".tf_wrapper.yml"):
+            if file.endswith(TF_WRAP_FILE):
                 wrapper_config_files.append(os.path.join(cur_path, file))
 
     return wrapper_config_files
@@ -110,10 +111,11 @@ def create_wrapper_config_obj(config_dir, wrapper_file=None):
     """
     if not wrapper_file:
         for file in os.listdir(config_dir):
-            if file.endswith(".tf_wrapper"):
+            if file.endswith(TF_WRAP_FILE):
                 wrapper_file = os.path.join(config_dir, file)
 
-    wrapper_config_obj: WrapperConfig = parse_wrapper_configs([wrapper_file])
+    wrapper_files = [wrapper_file] if wrapper_file else []
+    wrapper_config_obj: WrapperConfig = parse_wrapper_configs(wrapper_files)
     if wrapper_config_obj.depends_on:
         depends_on = []
         for dependency in wrapper_config_obj.depends_on:
@@ -139,7 +141,7 @@ def walk_and_graph_directory(starting_dir: str, config_dict) -> Tuple[networkx.D
     for root, _, files in os.walk(starting_dir):
         has_tf_wrapper = False
         for file in files:
-            if file.endswith(".tf_wrapper"):
+            if file.endswith(TF_WRAP_FILE):
                 has_tf_wrapper = True
                 wrapper_file = os.path.join(root, file)
                 wrapper_config_obj = create_wrapper_config_obj(root, wrapper_file)
@@ -171,12 +173,12 @@ def walk_without_graph_directory(starting_dir: str) -> List[str]:
     for root, _, files in os.walk(starting_dir):
         has_tf_wrapper = False
         for file in files:
-            if file.endswith(".tf_wrapper"):
+            if file.endswith(TF_WRAP_FILE):
                 has_tf_wrapper = True
                 wrapper_file = os.path.join(root, file)
                 wrapper_config_obj = create_wrapper_config_obj(root, wrapper_file)
                 if wrapper_config_obj.depends_on is not None:
-                    raise ValueError("Discovered dependency information")
+                    raise NoDependency("Discovered dependency information")
                 if not wrapper_config_obj.config:
                     continue
                 if not wrapper_config_obj.apply_automatically:
@@ -346,7 +348,7 @@ def parse_backend_config_for_dir(dir_path: str) -> Optional[BackendsConfig]:
     """
     has_tf_files = False
     for file_path in os.listdir(dir_path):
-        if '.terraform' in file_path or not file_path.endswith('tf'):
+        if '.terraform' in file_path or not file_path.endswith('.tf'):
             continue
 
         has_tf_files = True